requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gTTS>=2.4.0
boto3>=1.33.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
Pillow>=10.0.0
pypdfium2>=4.27.0
google-genai>=1.0.0
youtube-transcript-api>=0.6.0
supabase
celery[redis]>=5.3.0
redis>=5.0.0
sentry-sdk[fastapi,celery]
logtail-python